Meta-regression dataset helpers

Lightweight helpers to discover and load the curated meta-regression datasets shipped with this package (CSV files in `inst/extdata/metareg/`).

Details

During development (without installing the package), you can point these helpers at your local export folder via:

options(metahub.metareg_dir = "metahub/inst/derived/metareg")

Examples

if (FALSE) {
m <- metareg_manifest()
head(m)
}
m <- metareg_manifest()
head(m)
#>                      dataset_id source  source_pkg     source_object
#> 1       metadat_dat.lehmann2018   CRAN     metadat   dat.lehmann2018
#> 2       metadat_dat.mccurdy2020   CRAN     metadat   dat.mccurdy2020
#> 3 psymetadata_stasielowicz2019b   CRAN psymetadata stasielowicz2019b
#> 4 psymetadata_stasielowicz2019a   CRAN psymetadata stasielowicz2019a
#> 5  psymetadata_stasielowicz2020   CRAN psymetadata  stasielowicz2020
#> 6        metadat_dat.curtis1998   CRAN     metadat    dat.curtis1998
#>                                                                     source_title
#> 1                                  The Effect of Red on Perceived Attractiveness
#> 2                                               Studies on the Generation Effect
#> 3 Studies on the association between goal orientation and performance adaptation
#> 4 Studies on the association between goal orientation and performance adaptation
#> 5       Studies on the importance of cognitive ability in performance adaptation
#> 6              Studies on the Effects of Elevated CO2 Levels on Woody Plant Mass
#>   repo branch repo_license    k      measure n_mods
#> 1 <NA>   <NA>         CRAN   81 from_dataset     47
#> 2 <NA>   <NA>         CRAN 1653 from_dataset     25
#> 3 <NA>   <NA>         CRAN   86 from_dataset     22
#> 4 <NA>   <NA>         CRAN   76 from_dataset     21
#> 5 <NA>   <NA>         CRAN  133 from_dataset     21
#> 6 <NA>   <NA>         CRAN  102          SMD     20
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 moderators
#> 1 short_title|full_citation|year|prpublication|source_type|preregistered|gender|color_contrast|color_form|photo_type|photo_similarity|dv_type|dv_items|dv_scale|dv_scalebottom|dv_scaletop|location|continent|participants|participant_notes|design|eth_majority|eth_majority_detail|eth_stim|eth_match|red_age|control_age|color_red|color_control|red_original|color_match|presentation_control|stimuli_presentation|red_n|red_m|red_sd|control_n|control_m|control_sd|sd_diff|rm_r|control_attractiveness|notes|totalsamplesize|pooled|precision|abs_yi
#> 2                                                                                                                                                                                                                                                                                 article|experiment|sample|pairing|ni|stimuli|condition|gen_difficulty|manip_type|present_style|word_status|memory_test|memory_type|gen_constraint|learning_type|stimuli_relation|gen_mode|gen_task|attention|pacing|filler_task|age_grp|retention_delay|precision|abs_yi
#> 3                                                                                                                                                                                                                                                                                                                           study_id|es_id|author|pub_year|pub_type|peer_review|pgo_type|adapt_measures|adapt_method|adapt_method_specific|go_measure|financ_support|age|age_imputed|sex|perc_men|country|sample|level|complexity_dynamic|precision|abs_yi
#> 4                                                                                                                                                                                                                                                                                                                                    study_id|es_id|author|pub_year|pub_type|peer_review|adapt_measures|adapt_method|adapt_method_specific|go_measure|financ_support|age|age_imputed|sex|perc_men|country|sample|level|complexity_dynamic|precision|abs_yi
#> 5                                                                                                                                                                                                                                                                                         effect_id|author|pub_year|pub_type|peer_review|adapt_measures|adapt_method|adapt_method_specific|subj_adapt_definition|cog_abil_measure|ca_measure|financ_support|sex|men_prop|country|sample|task|complexity_component|complexity_coordinative|precision|abs_yi
#> 6                                                                                                                                                                                                                                                                                                                                                                                                          paper|genus|species|fungrp|co2ambi|co2elev|units|time|pot|method|stock|xtrt|level|n_ratio|sd_ratio|mean_ctrl|ntotal|log_ntotal|precision|abs_yi
#>          signature
#> 1 702e9bb2ca3a0247
#> 2 68c825d12beede6f
#> 3 23f03a8d8c2fc60a
#> 4 a36c3e28d1b33b84
#> 5 8e0550e6663727c8
#> 6 0675efb555cfaf8d